Fix logic for redirecting unsupported browsers to old editor
Browsers that support JavaScript, but that get served a no-JS
experience by MediaWiki (for example, IE 6), would always get
redirected to an URL with &venoscript=1 - even when it was unnecessary
because we're not in single-tab mode, and even when the form was
previously submitted, causing the loss of user input.

Follow-up to 188ec0f5d23c0daa55bdea1a019a6b9232fe411e.

Bug: T163226
Change-Id: I212baa0f4305559fa6f6abdfa230b76c122d4909