VisualEditorDataModule: Remove 'minoredit' and 'watchthis' messages
These messages are used for the checkboxes when saving the page.
But we don't actually use the data from this module; we get them
from the 'action=visualeditor' API in the 'checkboxesMessages'
property. This is necessary because MediaWiki extensions can add
extra checkboxes (e.g. FlaggedRevs), so a static list won't do.
These are unused since 393807462e9d04ec5e437cb50ef1d03e5644e9be
(2013) and it is confusing to keep them here.

Change-Id: Ia4ed2142292c3a22623f498c7f43ce590f167709