ve.init.MWVESwitchConfirmDialog: Fix gap above dropdown on MonoBook
In WikimediaUI theme, the height of the toolbar is defined with
'padding-top', but in Apex theme, it is defined with 'height'.
Set both properties to make this work with both themes.

Remove unused styles for the dropdown. It is now positioned by OOUI
(since T192505). This should have been done at the same time as
e1635fdc5225fbdcda941103c31382fa7456d1b4.

Bug: T194120
Change-Id: I9cac0c9458ae8cc26c5f056bb26686f8ad50c493