ve.init.mw.ArticleTarget: Fix error handling in #prepareCacheKey
If the paction=serializeforcache request fails, we were erroneously
converting it to a successful result with no value, which later causes
an exception, because since 381b58585cc4e6a74e64fb2669af2a7635b09100
other code expects the result to an an object.

The bug was introduced in 2015 in 07001001be794efba90fa0e81b212280db4b7784,
but until that recent change it would only cause a 'badcachekey'
error, which was handled correctly later.

Change-Id: Ie1ffc8c3e616a7d296f2186fb17eaf039971a44f